requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
asgiref==3.5.2
    # via django
attrs==22.1.0
    # via
    #   jsonschema
    #   pytest
certifi==2022.9.24
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
coreapi==2.3.3
    # via -r requirements.in
coreschema==0.0.4
    # via coreapi
cryptography==38.0.3
    # via
    #   jwcrypto
    #   python-jose
    #   social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
deprecated==1.2.13
    # via jwcrypto
django==4.1.2
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-oauth-toolkit
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-cors-headers==3.13.0
    # via -r requirements.in
django-oauth-toolkit==2.2.0
    # via drf-social-oauth2
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-social-oauth2
    #   drf-spectacular
djangorestframework-simplejwt==5.2.1
    # via -r requirements.in
drf-social-oauth2==1.2.1
    # via -r requirements.in
drf-spectacular==0.24.2
    # via -r requirements.in
ecdsa==0.18.0
    # via python-jose
idna==3.4
    # via requests
inflection==0.5.1
    # via drf-spectacular
iniconfig==1.1.1
    # via pytest
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
jsonschema==4.16.0
    # via drf-spectacular
jwcrypto==1.4.2
    # via django-oauth-toolkit
markupsafe==2.1.1
    # via jinja2
oauthlib==3.2.2
    # via
    #   django-oauth-toolkit
    #   requests-oauthlib
    #   social-auth-core
packaging==21.3
    # via pytest
pluggy==1.0.0
    # via pytest
psycopg2-binary==2.9.4
    # via -r requirements.in
py==1.11.0
    # via pytest
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.21
    # via cffi
pyjwt==2.5.0
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
pyparsing==3.0.9
    # via packaging
pyrsistent==0.18.1
    # via jsonschema
pytest==7.1.3
    # via
    #   -r requirements.in
    #   pytest-django
pytest-django==4.5.2
    # via -r requirements.in
python-jose[cryptography]==3.3.0
    # via drf-social-oauth2
python3-openid==3.2.0
    # via social-auth-core
pytz==2022.4
    # via djangorestframework
pyyaml==6.0
    # via drf-spectacular
requests==2.28.1
    # via
    #   coreapi
    #   django-oauth-toolkit
    #   requests-mock
    #   requests-oauthlib
    #   social-auth-core
requests-mock==1.10.0
    # via -r requirements.in
requests-oauthlib==1.3.1
    # via social-auth-core
rsa==4.9
    # via python-jose
six==1.16.0
    # via
    #   ecdsa
    #   requests-mock
social-auth-app-django==5.0.0
    # via drf-social-oauth2
social-auth-core==4.3.0
    # via social-auth-app-django
sqlparse==0.4.3
    # via django
tomli==2.0.1
    # via pytest
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-spectacular
urllib3==1.26.12
    # via requests
wrapt==1.14.1
    # via deprecated
